+++ src/server/game/Texts/CreatureTextMgr.cpp
@@ -331,7 +331,7 @@ void CreatureTextMgr::SendSound(Creature* source, uint32 sound, ChatMsg msgType,
 
     WorldPacket data(SMSG_PLAY_SOUND, 4 + 8);
     data << uint32(sound);
-    data << uint64(source->GetGUID());
+    data << source->GetGUID();
     SendNonChatPacket(source, &data, msgType, whisperTarget, range, team, gmOnly);
 }
